+++ CHANGELOG.md
@@ -11,6 +11,8 @@ maintenance mode.
 
 ### Changed
 
+- concord-server: fixed a bug preventing relative file upload paths
+from working with `/api/v1/process` endpoint;
 - concord-server: the session cookie (`JSESSIONID`) is now marked as
 `HttpOnly`;
 - concord-common: ensure that `CONCORD_TMP_DIR` environment variable